updated model and api client
