+++ NYTBestBooks/Models/Book.swift
@@ -8,6 +8,42 @@
 
 import UIKit
 
-struct Book: Codable & Equatable {
-    
+struct SearchResult: Codable & Equatable {
+    let results: SearchInfo
 }
+
+struct SearchInfo: Codable & Equatable {
+    let listName: String
+    let books: [Book]
+    
+    private enum CodingKeys: String, CodingKey {
+    case listName = "list_name"
+    case books
+    }
+}
+
+struct Book: Codable & Equatable {
+    let rank: Double
+    let description: String
+    let title: String
+    let author: String
+    let bookImage: String
+    let bookLinks: [BuyLink]
+    let weeksOnList: Double
+    
+    private enum CodingKeys: String, CodingKey {
+        case rank
+        case description
+        case title
+        case author
+        case bookImage = "book_image"
+        case bookLinks = "buy_links"
+        case weeksOnList = "weeks_on_list"
+    }
+}
+
+struct BuyLink: Codable & Equatable {
+    let name: String
+    let url: String
+}
+
